requirements.txt
fastapi
pandas
python-dotenv
nicegui